wip: basic refactor to slot and slot_iterator
