+++ src/models/slot/iterator.rs
@@ -1,145 +1,6 @@
-use crate::models::goal::{Day, TimeFilter};
 use crate::models::repetition::Repetition;
-use crate::models::slot::Slot;
-use chrono::prelude::*;
-use chrono::Days;
-use chrono::Duration;
-use chrono::NaiveDateTime;
 
-// derive Debug for TimeSlotsIterator
-#[derive(Debug)]
-/// An iterator that returns slots that conform to a repetition,
-/// with optional filters (after/before/Mondays/Weekdays/Weekends/Bank holidays...)
-/// e.g. iterate over all MONDAYS between 1st September 2022 and 30th September 2022.
-/// e.g. iterate over all DAYS between 1st September 2022 and 30th September 2022.
-/// e.g. iterate over all HOURS between 1st September 2022 and 30th September 2022.
-/// e.g. iterate over all 'every two hours' slots between times 10-21 for time period X-Y
-pub(crate) struct TimeSlotsIterator {
-    timeline: Vec<Slot>,
-    repetition: Option<Repetition>,
-    filters: Option<TimeFilter>,
-    current_start_position: NaiveDateTime,
-}
-
-impl TimeSlotsIterator {
-    pub fn new(
-        start: NaiveDateTime,
-        end: NaiveDateTime,
-        repetition: Option<Repetition>,
-        filters: Option<TimeFilter>,
-    ) -> TimeSlotsIterator {
-        let mut result = TimeSlotsIterator {
-            timeline: vec![Slot { start, end }],
-            repetition,
-            filters,
-            current_start_position: start, //override after applying filters
-        };
-        result.apply_filters();
-        result.current_start_position = result.timeline[0].start;
-        result
-    }
-
-    fn apply_filters(&mut self) {
-        match &self.filters {
-            Some(time_filter) => {
-                if time_filter.after_time.is_some() || time_filter.before_time.is_some() {
-                    let mut result: Vec<Slot> = vec![];
-                    for slot in self.timeline.iter_mut() {
-                        let mut daily_slots = slot.divide_in_days();
-                        if time_filter.after_time.is_some() && time_filter.before_time.is_some() {
-                            //after and before time
-                            for daily_slot in daily_slots.iter_mut() {
-                                let before_datetime = daily_slot
-                                    .start
-                                    .with_hour(time_filter.before_time.unwrap() as u32)
-                                    .unwrap();
-                                let after_datetime = daily_slot
-                                    .start
-                                    .with_hour(time_filter.after_time.unwrap() as u32)
-                                    .unwrap();
-
-                                if after_datetime > before_datetime {
-                                    if daily_slot.start < before_datetime {
-                                        result.push({
-                                            Slot {
-                                                start: daily_slot.start,
-                                                end: before_datetime,
-                                            }
-                                        });
-                                    }
-                                    if daily_slot.end > after_datetime {
-                                        result.push({
-                                            Slot {
-                                                start: after_datetime,
-                                                end: daily_slot.end,
-                                            }
-                                        });
-                                    }
-                                } else {
-                                    result.push({
-                                        Slot {
-                                            start: after_datetime,
-                                            end: before_datetime,
-                                        }
-                                    })
-                                }
-                            }
-                        }
-                        if time_filter.after_time.is_some() && time_filter.before_time.is_none() {
-                            //only after time
-                            for daily_slot in daily_slots.iter_mut() {
-                                let after_datetime = daily_slot
-                                    .start
-                                    .with_hour(time_filter.after_time.unwrap() as u32)
-                                    .unwrap();
-                                if daily_slot.start < after_datetime {
-                                    daily_slot.start = after_datetime;
-                                }
-                                result.push(*daily_slot);
-                            }
-                        }
-                        if time_filter.after_time.is_none() && time_filter.before_time.is_some() {
-                            //only before time
-                            for daily_slot in daily_slots.iter_mut() {
-                                let before_datetime = daily_slot
-                                    .start
-                                    .with_hour(time_filter.before_time.unwrap() as u32)
-                                    .unwrap();
-                                if daily_slot.end > before_datetime {
-                                    daily_slot.end = before_datetime;
-                                }
-                                result.push(*daily_slot);
-                            }
-                        }
-                    }
-                    self.timeline = result;
-                }
-
-                match &time_filter.on_days {
-                    Some(on_days) => {
-                        let mut result: Vec<Slot> = vec![];
-                        for slot in self.timeline.iter_mut() {
-                            let mut daily_slots = slot.divide_in_days();
-
-                            for daily_slot in daily_slots.iter_mut() {
-                                // Check if the weekday matches with the given on days filter value
-                                //  and push it to result vector if true.
-                                let start_day: String = slot.start.weekday().to_string();
-                                if on_days.contains(&Day::from(start_day)) {
-                                    result.push(*daily_slot);
-                                }
-                            }
-                        }
-
-                        self.timeline = result;
-                    }
-                    None => (),
-                }
-            }
-            None => (),
-        }
-    }
-}
+use super::TimeSlotsIterator;
 
 impl Iterator for TimeSlotsIterator {
     type Item = Vec<Slot>;
